+++ CHANGELOG.md
@@ -5,8 +5,10 @@
 ### Added
 
 - clap example
+- support timestamps with microsecond and nanosecond granularity
 
 ### Changed
 
 - improved performance (https://github.com/cardoe/stderrlog-rs/pull/2)
 - improved docopt example in README
+- migrated from time to chrono